M129 instructor view

+++ StudentDriver/StudentDriver/Models/DrivePoint.cs
@@ -12,6 +12,6 @@ namespace StudentDriver.Models
         public double Latitude { get; set; }
         public double Longitude { get; set; }
         public DateTime PointDateTime { get; set; }
-        public float Speed { get; set; }
+        public double Speed { get; set; }
     }
 }
